fix undefined macros
The worst of them were an undefined UINTPTR_MAX in RM_PLATFORM_*, which
could both be false and caused a stat struct to be mis-casted in
traverse.c, and a non-macro HASHER_FADVISE_FLAGS that made
rm_hasher_request_readahead a no-op since commit 31cd32f1.

Also add a static assert in the usual ADD_FILE to make sure it never
casts between incompatible stat structs, and -Werror=undef so we don't
allow undefined macros to silently evaluate to zero.

The UINTPTR_MAX issue is a regression caused by 90edf021, which removed
the inttypes.h include in config.h.in.

Fixes #549